refactor create Idenitity Screen
